wip: migrate to apollo-modules-node and backend-modules
move all auth and user related graphql queries and resolvers
and middleware to backend-modules-auth.
move document graphql queries and resolvers to backend-module-documents.
require local graphql with apollo-modules-node
don't yet depend backend-modules-base, which handles all the express, graphql and
database heavy lifting, but use a copy of it's code locally.

TODOs
- [  ] migrate test setup
- solve npm import errors
