--- graphql/resolvers/RootQuerys/me.js
@@ -1,3 +1,0 @@
-module.exports = async (_, args, { user }) => {
-  return user
-}
