--- graphql/resolvers/RootSubscription/index.js
@@ -1,5 +1,0 @@
-require('fs').readdirSync(__dirname).forEach(function (file) {
-  if (file !== 'index.js') {
-    module.exports[file.split('.')[0]] = require('./' + file)
-  }
-})
