+++ graphql/resolvers/_mutations/commit.js
@@ -3,7 +3,7 @@ const visit = require('unist-util-visit')
 const dataUriToBuffer = require('data-uri-to-buffer')
 const MDAST = require('../../../lib/mdast/mdast')
 const { unprefixUrl } = require('../../../lib/assets')
-const { ensureUserHasRole } = require('../../../lib/Roles')
+const { Roles: { ensureUserHasRole } } = require('backend-modules-auth')
 const superb = require('superb')
 const superheroes = require('superheroes')
 const sleep = require('await-sleep')
