+++ graphql/resolvers/_mutations/uncommittedChanges.js
@@ -1,4 +1,4 @@
-const { ensureUserHasRole } = require('../../../lib/Roles')
+const { Roles: { ensureUserHasRole } } = require('backend-modules-auth')
 
 module.exports = async (
   _,
