+++ lib/t.js
@@ -1,4 +1,3 @@
-const { getFormatter } = require('./translate')
+const { getFormatter } = require('backend-modules-translate')
 const MESSAGES = require('./translations.json').data
-
 module.exports = getFormatter(MESSAGES)
