+++ package.json
@@ -9,11 +9,16 @@
     "@project-r/template-newsletter": "^1.0.0",
     "apollo-engine": "^0.5.2",
     "apollo-fetch": "^0.6.0",
+    "apollo-modules-node": "file:../apollo-modules-node",
     "apollo-server-express": "^1.2.0",
     "await-sleep": "0.0.1",
+    "backend-modules-auth": "file:../backend-modules/packages/auth",
+    "backend-modules-documents": "file:../backend-modules/packages/documents",
+    "backend-modules-mail": "file:../backend-modules/packages/mail",
+    "backend-modules-translate": "file:../backend-modules/packages/translate",
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.2",
-    "connect-pg-simple": "^3.1.2",
+    "check-env": "^1.3.0",
     "cookie-parser": "^1.4.3",
     "cors": "^2.8.4",
     "cron": "^1.3.0",
@@ -24,27 +29,22 @@
     "db-migrate-pg": "^0.2.5",
     "deep-diff": "^0.3.8",
     "dotenv": "^4.0.0",
-    "email-validator": "^1.1.1",
+    "export-files": "^2.1.1",
     "express": "^4.16.2",
-    "express-session": "^1.15.6",
     "file-type": "^7.2.0",
-    "geoip-database": "^1.0.9",
     "github": "^12.0.1",
     "graphql": "^0.11.7",
     "graphql-redis-subscriptions": "^1.3.1",
     "graphql-subscriptions": "^0.5.4",
     "graphql-tag": "^2.5.0",
-    "graphql-tools": "^2.7.2",
+    "graphql-tools": "github:orbiting/graphql-tools",
     "graphql-type-json": "^0.1.4",
     "immutable": "^3.8.2",
     "isomorphic-unfetch": "^2.0.0",
     "jsonwebtoken": "^8.1.0",
-    "kraut": "^0.1.1",
     "leaked-handles": "^5.2.0",
     "lodash": "^4.17.4",
-    "maxmind": "^2.3.0",
     "mdast-react-render": "^0.1.1",
-    "passport": "^0.4.0",
     "pogi": "^2.4.0",
     "query-string": "^5.0.1",
     "react": "^16.0.0",
